+++ lib/domain/repositories/note/note_repository.dart
@@ -1,12 +1,17 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:primary_school/domain/models/note_model/note_model.dart';
 
 class NoteRepository {
   Stream<List<NoteModel>> getNotesStream() {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     return FirebaseFirestore.instance
         .collection('users')
-        .doc("N4idavG5HPeXOMj0MGYHYp2Pvzt1")
+        .doc(userId)
         .collection('noteItems')
         .snapshots()
         .map((querySnapshot) {
@@ -40,9 +45,13 @@ class NoteRepository {
     DateTime updatedDate,
     Color selectedColor,
   ) async {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     await FirebaseFirestore.instance
         .collection('users')
-        .doc("N4idavG5HPeXOMj0MGYHYp2Pvzt1")
+        .doc(userId)
         .collection('noteItems')
         .add({
       'title': title,
@@ -61,9 +70,13 @@ class NoteRepository {
     DateTime updatedDate,
     String docId,
   ) async {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     await FirebaseFirestore.instance
         .collection('users')
-        .doc("N4idavG5HPeXOMj0MGYHYp2Pvzt1")
+        .doc(userId)
         .collection('noteItems')
         .doc(docId)
         .update({
@@ -75,9 +88,13 @@ class NoteRepository {
   }
 
   Future<void> delete({required String id}) async {
+    final userId = FirebaseAuth.instance.currentUser?.uid;
+    if (userId == null) {
+      throw Exception('User is not logged in');
+    }
     await FirebaseFirestore.instance
         .collection('users')
-        .doc("N4idavG5HPeXOMj0MGYHYp2Pvzt1")
+        .doc(userId)
         .collection('noteItems')
         .doc(id)
         .delete();
